refactor(codemod): rename `hook-call-transformer` to `use-query-like` transformer
I think this name describes better the purpose of this transformer.
